requirements.txt
discord.py
feedparser
python-dotenv
google-genai
requests
